Update the `natmanager` package itself.

Update the `natmanager` package itself.

selfupdate(
  source = c("GITHUB", "CRAN"),
  upgrade.dependencies = "always",
  force = FALSE,
  ...
)

Arguments

source
Location from which to obtain a newer version of natmanager. Defaults to GITHUB since this may well have a newer version than the CRAN package repository.

upgrade.dependencies
Whether to install dependencies of natmanager. See the upgrade argument of install_github for details. The default will go ahead and always do this is as necessary.

force
Force self update even if there doesn't seem to be an update (default FALSE)

...
extra arguments to pass to install_github or install_cran.

Value

Logical indicating whether an update was required (invisibly).

See also

Examples

if (FALSE) {
natmanager::selfupdate()
}